test(discord): split retry coverage
